--- .gitattributes
@@ -1,22 +1,0 @@
-# Set the default behavior, in case people don't have core.autocrlf set.
-* text=auto
-
-# convert to lf on checkout.
-*.go text eol=lf
-*.mod text eol=lf
-*.sum text eol=lf
-*.md text eol=lf
-*.svg text eol=lf
-*.http text eol=lf
-*.ini text eol=lf
-*.conf text eol=lf
-*.sh text eol=lf
-*.yaml text eol=lf
-*.proto text eol=lf
-*.json text eol=lf
-Dockerfile text eol=lf
-
-# Denote all files that are truly binary and should not be modified.
-*.png binary
-*.jpg binary
-*.jpeg binary
