+++ .github/workflows/test.yml
@@ -9,8 +9,14 @@ jobs:
   test-project:
     runs-on: ubuntu-20.04
     steps:
+      - name: Setup
+        uses: actions/setup-go@v4
+        with:
+          go-version: "1.21"
+      - run: go version
+
       - name: Checkout
-        uses: actions/checkout@v3
+        uses: actions/checkout@v4
 
       - name: Test
         run: make test
