+++ HISTORY.md
@@ -1,5 +1,11 @@
 ## ✒ 历史版本的特性介绍 (Features in old versions)
 
+### v0.8.0
+
+> 此版本发布于 2024-08-09
+
+* 重构，调整使用姿势
+
 ### v0.5.2
 
 > 此版本发布于 2023-07-27
