+++ errors.go
@@ -1,4 +1,4 @@
-// Copyright 2023 FishGoddess. All rights reserved.
+// Copyright 2024 FishGoddess. All rights reserved.
 // Use of this source code is governed by a MIT style
 // license that can be found in the LICENSE file.
 
@@ -6,15 +6,28 @@ package errors
 
 import (
 	"errors"
-	"fmt"
 )
 
-// New returns a string error.
 func New(text string) error {
 	return errors.New(text)
 }
 
-// Newf returns a string error formatted with params.
-func Newf(text string, params ...interface{}) error {
-	return fmt.Errorf(text, params...)
+// Is is a shortcut of errors.Is.
+func Is(err, target error) bool {
+	return errors.Is(err, target)
+}
+
+// As is a shortcut of errors.As.
+func As(err error, target any) bool {
+	return errors.As(err, target)
+}
+
+// Unwrap is a shortcut of errors.Unwrap.
+func Unwrap(err error) error {
+	return errors.Unwrap(err)
+}
+
+// Join is a shortcut of errors.Join.
+func Join(errs ...error) error {
+	return errors.Join(errs...)
 }
